+++ src/_root/_app-sidebar.tsx
@@ -2,7 +2,7 @@ import AppSidebar from '@/_root/components/app-sidebar'
 import { SidebarLink } from '@/components/types'
 import { SidebarInset, SidebarProvider, SidebarTrigger } from '@/components/ui/sidebar'
 import { useAuthContext } from '@/context/AuthProvider'
-import { getChats } from '@/lib/api/recipes'
+import { getChats } from '@/lib/api/llm_chats'
 import { useQuery } from '@tanstack/react-query'
 import { Archive, CookingPot, House, Pencil } from 'lucide-react'
 import { useEffect, useState } from 'react'
